NX-507: Exploded more IT support classes
